refactor utils into folder
