+++ electron/utils/deepFreeze.spec.ts
@@ -1,4 +1,4 @@
-import { deepFreeze } from './utils';
+import { deepFreeze } from './deepFreeze';
 
 describe('deepFreeze', () => {
   it('should freeze all properties', () => {
