+++ electron/utils/deepFreeze.ts
