+++ electron/utils/index.ts
@@ -1,0 +1,1 @@
+export { deepFreeze } from './deepFreeze';
